cli: add a new jaeger cmd
Adds a new jaeger high level folder which contains the cli code. This
contains its own `Values` based on `jaeger/values.yaml`. It re-uses a
lot of code from cli install, and follows the code format and logic.

A lot of code is copy pasted, and could have a refactor commit,
to make it re-usable.

Signed-off-by: GitHub <noreply@github.com>